Use `nx generate` instead of short alias
`nx generate` is used in other places in the document. We should use `nx generate` everywhere for consistency, and also because it is more descriptive to readers who are not familiar with Nx command aliases.

Additionally, a sentence referring to the plugin version has been removed as the snippet it refers to does not exist.